refactoring: FindDuo
refactoring: FindDuo
