+++ src/components/Duo/Common/Modal/Modal.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Modal = () => {
+  return <div>Modal</div>;
+};
+
+export default Modal;
