+++ src/components/Duo/Duo.tsx
@@ -1,10 +1,11 @@
-import { useCallback, useEffect, useState } from 'react';
+import { ChangeEvent, useCallback, useEffect, useRef, useState } from 'react';
 import WriteDuo from './WriteDuo/WriteDuo';
 import ListDuo from './ListDuo/ListDuo';
 import styles from './Duo.module.scss';
 
 const Duo = () => {
   const [isModal, setIsModal] = useState<boolean>(false);
+  const modalRef = useRef<HTMLFormElement>(null);
 
   const onToggleClick = useCallback(() => setIsModal((e) => !e), [isModal]);
 
@@ -20,6 +21,19 @@ const Duo = () => {
     onScrollToggle();
   });
 
+  useEffect(() => {
+    const clickOutside = (e: any) => {
+      if (isModal && !modalRef.current?.contains(e.target)) {
+        setIsModal(false);
+      }
+      document.addEventListener('mousedown', clickOutside);
+      return () => {
+        document.removeEventListener('mousedown', clickOutside);
+      };
+    };
+  }, [isModal]);
+  // console.log(modalRef);
+
   return (
     <>
       <article className={styles.duo_article}>
@@ -28,7 +42,7 @@ const Duo = () => {
             소환사 등록하기
           </button>
         </div>
-        {isModal && <WriteDuo onToggleClick={onToggleClick} />}
+        {isModal && <WriteDuo onToggleClick={onToggleClick} modalRef={modalRef} />}
         <ListDuo />
       </article>
     </>
