+++ src/components/Duo/ListDuo/DuoCards/DuoCards.tsx
@@ -1,6 +1,6 @@
 import { Link } from 'react-router-dom';
-import { DuoType, LaneType, TierType } from '../utils/DuoType';
-import { positions, tiers } from '../utils/DuoArr';
+import { DuoType, LaneType, TierType } from '../../utils/DuoType';
+import { positions, tiers } from '../../utils/DuoArr';
 import { LANE_ICONS_URL, TIER_IMG_URL } from '@/components/Duo/Constants/constant';
 import styles from './DuoCards.module.scss';
