+++ src/components/Duo/ListDuo/DuoInfo/DuoInfo.module.scss
@@ -1,4 +1,4 @@
-@import '../common.scss';
+@import '../../common.scss';
 
 /* DuoInfo.module.scss */
