+++ src/components/Duo/WriteDuo/WriteDuo.module.scss
@@ -22,6 +22,8 @@ $bgColor-2: #282830;
 }
 
 .wrapper {
+  width: auto;
+  height: auto;
   display: flex;
   justify-content: center;
   align-items: center;
