+++ src/route.tsx
@@ -6,7 +6,7 @@ import Duo from '@/pages/Duo';
 import LCK from '@/pages/LCK';
 import NotFound from '@/pages/NotFound';
 
-import DuoInfo from './components/Duo/DuoInfo/DuoInfo';
+import DuoInfo from './components/Duo/ListDuo/DuoInfo/DuoInfo';
 
 import Team from '@/components/LCK/components/Team/team_main/Team';
 import Match from '@/components/LCK/components/Match/Match';
